ref(workspace): move and rename workspace
