+++ src/workspace/repository_tests.rs
@@ -1,5 +1,5 @@
 use super::super::common_test;
-use crate::workspace::Workspace;
+use crate::workspace::repository::Repository;
 
 #[cfg(test)]
 mod latest_tag {
@@ -8,7 +8,7 @@ mod latest_tag {
     #[test]
     fn no_tag_set() -> Result<(), git2::Error> {
         let (dir, _) = common_test::repo_init();
-        let ws = Workspace::new(dir)?;
+        let ws = Repository::new(dir)?;
         let tag = ws.latest_tag();
         assert!(tag.is_err());
         Ok(())
@@ -18,13 +18,13 @@ mod latest_tag {
     fn one_tag() -> Result<(), git2::Error> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        let (commit, _) = common_test::commit(&repo, "file");
+        let (commit, _) = common_test::commit(&repo, "file", None);
         let obj = repo.find_object(commit, None)?;
         let sig = repo.signature()?;
         let tag_name = "v0.6.6.6";
         repo.tag(tag_name, &obj, &sig, "msg", false)?;
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.latest_tag()?;
         assert_eq!(tag_name, tag);
         Ok(())
@@ -39,7 +39,7 @@ mod latest_tag {
         let tag_name = "tag2";
         common_test::commit_tag(&repo, "file", tag_name);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.latest_tag()?;
         assert_eq!(tag_name, tag);
         Ok(())
@@ -57,7 +57,7 @@ mod previous_tag {
         let tag = "tag";
         common_test::commit_tag(&repo, "file", tag);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.validate_tag("not_exists");
         assert!(tag.is_err());
         Ok(())
@@ -70,7 +70,7 @@ mod previous_tag {
         let tag = "tag";
         let (_, oid) = common_test::commit_tag(&repo, "file", tag);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.validate_tag(tag)?;
         assert_eq!(oid, tag);
         Ok(())
@@ -83,7 +83,7 @@ mod previous_tag {
         let tag = "tag";
         common_test::commit_lightweight_tag(&repo, "file", tag);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.validate_tag("not_exists");
         assert!(tag.is_err());
         Ok(())
@@ -96,7 +96,7 @@ mod previous_tag {
         let tag = "tag";
         let (_, oid) = common_test::commit_lightweight_tag(&repo, "file", tag);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.validate_tag(tag)?;
         assert_eq!(oid, tag);
         Ok(())
@@ -108,7 +108,7 @@ mod previous_tag {
         let repo = git2::Repository::open(&dir)?;
         common_test::commit_tag(&repo, "file", "tag");
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.previous_tag("not_exists");
         assert!(tag.is_err());
         Ok(())
@@ -121,7 +121,7 @@ mod previous_tag {
         let tag_name = "tag1";
         let (commit, _) = common_test::commit_tag(&repo, "file", tag_name);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.previous_tag(tag_name)?;
         assert_eq!(commit.to_string(), tag);
         Ok(())
@@ -131,21 +131,21 @@ mod previous_tag {
     fn commit_in_the_middle() -> Result<(), Box<dyn std::error::Error>> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        common_test::commit(&repo, "file1");
+        common_test::commit(&repo, "file1", None);
 
         let tag1 = "tag1";
         common_test::commit_tag(&repo, "file2", tag1);
 
-        common_test::commit(&repo, "file3");
-        common_test::commit(&repo, "file4");
+        common_test::commit(&repo, "file3", None);
+        common_test::commit(&repo, "file4", None);
 
         let tag2 = "tag2";
         common_test::commit_tag(&repo, "file5", tag2);
 
-        common_test::commit(&repo, "file5");
-        common_test::commit(&repo, "file6");
+        common_test::commit(&repo, "file5", None);
+        common_test::commit(&repo, "file6", None);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let tag = ws.previous_tag(tag2)?;
         assert_eq!(tag1.to_string(), tag);
         Ok(())
@@ -160,12 +160,12 @@ mod commits_between_tags {
     fn from_tag_not_in_repo() -> Result<(), Box<dyn std::error::Error>> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        common_test::commit(&repo, "file1");
+        common_test::commit(&repo, "file1", None);
 
         let tag = "tag";
         common_test::commit_tag(&repo, "file2", tag);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let res = ws.commits_between_tags("tag_not_found", tag);
         assert!(res.is_err());
 
@@ -176,12 +176,12 @@ mod commits_between_tags {
     fn to_tag_not_in_repo() -> Result<(), Box<dyn std::error::Error>> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        common_test::commit(&repo, "file1");
+        common_test::commit(&repo, "file1", None);
 
         let tag = "tag";
         common_test::commit_tag(&repo, "file2", tag);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let res = ws.commits_between_tags(tag, "tag_not_found");
         assert!(res.is_err());
 
@@ -192,9 +192,9 @@ mod commits_between_tags {
     fn from_and_to_point_to_same_commit() -> Result<(), Box<dyn std::error::Error>> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        common_test::commit(&repo, "file1");
+        common_test::commit(&repo, "file1", None);
 
-        let (commit, _) = common_test::commit(&repo, "file2");
+        let (commit, _) = common_test::commit(&repo, "file2", None);
         let obj = repo.find_object(commit, None)?;
         let sig = repo.signature()?;
         let tag1 = "tag1";
@@ -202,7 +202,7 @@ mod commits_between_tags {
         let tag2 = "tag2";
         repo.tag(tag2, &obj, &sig, "msg", false)?;
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let res = ws.commits_between_tags(tag1, tag2);
         assert!(res.is_err());
 
@@ -213,7 +213,7 @@ mod commits_between_tags {
     fn adjacent_commits() -> Result<(), Box<dyn std::error::Error>> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        common_test::commit(&repo, "file1");
+        common_test::commit(&repo, "file1", None);
 
         let tag1 = "tag1";
         common_test::commit_tag(&repo, "file2", tag1);
@@ -221,7 +221,7 @@ mod commits_between_tags {
         let tag2 = "tag2";
         let (commit2, _) = common_test::commit_tag(&repo, "file3", tag2);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let res: Vec<git2::Oid> = ws
             .commits_between_tags(tag1, tag2)?
             .iter()
@@ -236,22 +236,22 @@ mod commits_between_tags {
     fn multiple_commits() -> Result<(), Box<dyn std::error::Error>> {
         let (dir, _) = common_test::repo_init();
         let repo = git2::Repository::open(&dir)?;
-        common_test::commit(&repo, "file_p1");
-        common_test::commit(&repo, "file_p2");
+        common_test::commit(&repo, "file_p1", None);
+        common_test::commit(&repo, "file_p2", None);
 
         let tag1 = "tag1";
         common_test::commit_tag(&repo, "file2", tag1);
 
-        let (commit2, _) = common_test::commit(&repo, "file3");
-        let (commit3, _) = common_test::commit(&repo, "file4");
+        let (commit2, _) = common_test::commit(&repo, "file3", None);
+        let (commit3, _) = common_test::commit(&repo, "file4", None);
 
         let tag2 = "tag2";
         let (commit4, _) = common_test::commit_tag(&repo, "file5", tag2);
 
-        common_test::commit(&repo, "file5");
-        common_test::commit(&repo, "file6");
+        common_test::commit(&repo, "file5", None);
+        common_test::commit(&repo, "file6", None);
 
-        let ws = Workspace::new(&dir)?;
+        let ws = Repository::new(&dir)?;
         let res: Vec<git2::Oid> = ws
             .commits_between_tags(tag1, tag2)?
             .iter()
